static, dynamic and partial prerendering(ppr)
