+++ next.config.mjs
@@ -1,5 +1,9 @@
 /** @type {import('next').NextConfig} */
 
-const nextConfig = {};
+const nextConfig = {
+  experimental: {
+    ppr: "incremental",
+  }
+};
 
 export default nextConfig;
